requirements.txt
opencv-python
ultralytics
rplidar-roboticia
numpy
matplotlib
streamlit
scipy